+++ src/Managers/Assistants/IndexAssistant.php
@@ -56,9 +56,6 @@ trait IndexAssistant
         return true;
     }
 
-    /**
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
-     */
     public function index(Request $request)
     {
         $this->guard('index');
@@ -70,68 +67,68 @@ trait IndexAssistant
         View::share('model', $model = $this->managedModelClassInstance());
 
         return $this->resource->getIndexView();
-
-        if ($model instanceof Nestable) {
-            // TODO: this should be changed to the repository pattern like:
-            // app($resource->indexRepository(), ['resourceKey' => $resourceKey])->applyFilters(request()->all())->getNestableResults()
-            // indexModelIds can then be removed
-            $filteredModelIds = $this->indexModelIds();
-
-            $filteredTree = $this
-                ->getTree()
-                ->shake(fn ($node) => in_array($node->getModel()->getKey(), $filteredModelIds));
-
-            View::share('tree', $filteredTree);
-            View::share('originalModels', PairOptions::toPairs($this->getTree()->pluck('id', fn (NestedNode $nestedNode) => $nestedNode->getBreadCrumbLabel())));
-        } else {
-            // Used for duplicate action
-            View::share('originalModels', ModelReferencePresenter::toSelectValues($model::all(), false, false));
-            View::share('models', $this->indexModels());
-        }
+//
+//        if ($model instanceof Nestable) {
+//            // TODO: this should be changed to the repository pattern like:
+//            // app($resource->indexRepository(), ['resourceKey' => $resourceKey])->applyFilters(request()->all())->getNestableResults()
+//            // indexModelIds can then be removed
+//            $filteredModelIds = $this->indexModelIds();
+//
+//            $filteredTree = $this
+//                ->getTree()
+//                ->shake(fn ($node) => in_array($node->getModel()->getKey(), $filteredModelIds));
+//
+//            View::share('tree', $filteredTree);
+//            View::share('originalModels', PairOptions::toPairs($this->getTree()->pluck('id', fn (NestedNode $nestedNode) => $nestedNode->getBreadCrumbLabel())));
+//        } else {
+//            // Used for duplicate action
+//            View::share('originalModels', ModelReferencePresenter::toSelectValues($model::all(), false, false));
+//            View::share('models', $this->indexModels());
+//        }
 
         return $this->resource->getIndexView();
     }
 
-    private function getTree(): iterable
-    {
-        // TODO: work with the indexRepository
-        // app($resource->indexRepository(), ['resourceKey' => $resourceKey])->applyFilters(request()->all())->getNestableResults()
+//    private function getTree(): iterable
+//    {
+//        // TODO: work with the indexRepository
+//        // app($resource->indexRepository(), ['resourceKey' => $resourceKey])->applyFilters(request()->all())->getNestableResults()
+//
+//        return app(NestableRepository::class)->getTree($this->resource::resourceKey());
+//    }
 
-        return app(NestableRepository::class)->getTree($this->resource::resourceKey());
-    }
-
-    protected function indexModels(): Paginator
-    {
-        // Apply filtering - this also includes default sorting
-        $this->filters()->apply($builder = $this->managedModelClass()::query());
-
-        if ($this->managedModelClassInstance() instanceof Visitable) {
-            $builder->with(['urls']);
-        }
-
-        if ($this->managedModelClassInstance() instanceof Taggable) {
-            $builder->with(['tags']);
-        }
-
-        if (! $pagination = $this->resource->getIndexPagination()) {
-            return $builder->get();
-        }
-
-        return $builder->paginate($pagination)->onEachSide(1)->withQueryString();
-    }
-
-    protected function indexModelIds(): array
-    {
-        $this->filters()->apply($builder = $this->managedModelClass()::query());
-
-        $idColumn = $this->managedModelClassInstance()->getKeyName();
-
-        return $builder
-            ->select($idColumn)
-            ->get()
-            ->pluck($idColumn)
-            ->toArray();
-    }
+//    protected function indexModels(): Paginator
+//    {
+//        // Apply filtering - this also includes default sorting
+//        $this->filters()->apply($builder = $this->managedModelClass()::query());
+//
+//        if ($this->managedModelClassInstance() instanceof Visitable) {
+//            $builder->with(['urls']);
+//        }
+//
+//        if ($this->managedModelClassInstance() instanceof Taggable) {
+//            $builder->with(['tags']);
+//        }
+//
+//        if (! $pagination = $this->resource->getIndexPagination()) {
+//            return $builder->get();
+//        }
+//
+//        return $builder->paginate($pagination)->onEachSide(1)->withQueryString();
+//    }
+//
+//    protected function indexModelIds(): array
+//    {
+//        $this->filters()->apply($builder = $this->managedModelClass()::query());
+//
+//        $idColumn = $this->managedModelClassInstance()->getKeyName();
+//
+//        return $builder
+//            ->select($idColumn)
+//            ->get()
+//            ->pluck($idColumn)
+//            ->toArray();
+//    }
 
     public function filters(): Filters
     {