+++ src/TableNew/Livewire/TableComponent.php
@@ -84,6 +84,10 @@ class TableComponent extends Component
         if ($this->getTable()->hasQuery()) {
             $builder = $this->getTable()->getQuery()();
 
+            foreach($this->getTable()->getAddedQueries() as $addedQuery) {
+                $addedQuery($builder);
+            }
+
             $this->applyQueryFilters($builder);
             $this->applyQuerySorters($builder);